// src/fireflies/types.ts
export type Firefly = {
    x: number;
    y: number;
    size: number;
    phase: number;
    glowSpeed: number;
    freqX1: number;
    freqX2: number;
    freqY1: number;
    freqY2: number;
    phaseX1: number;
    phaseX2: number;
    phaseY1: number;
    phaseY2: number;
    amplitudeX: number;
    amplitudeY: number;
};